fix(engine): clamp ranking size tiebreaker + integer permille profile fields (2 CRITICAL review findings)
From the adversarial review of the engine crates.

ranking: a negative video_size (i64) drove the size tiebreaker ~1.3e9 below the junk floor, letting an add-on self-suppress its streams. Clamp the contribution to [0, 12]. Regression test added.

state: text_scale (ProfileSettings) and progress (CwItem) were f32. The roster merge key serializes the profile, and floats serialize differently across Rust/TS/Swift (ryu 1.0 vs 1), which would break CRDT convergence cross-platform. Both are now u32 permille (1000 = 1.0x).

Whole workspace stays green; clippy -D warnings clean.

+++ vortx-core/crates/state/src/library.rs
@@ -75,12 +75,14 @@ pub struct HistoryEntry {
 }
 
 /// A Continue Watching rail entry (emitted by the engine for the active profile).
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct CwItem {
     pub id: String,
     pub name: String,
+    /// Playback progress in permille (0..=1000). An integer (not a float) so per-profile sync is
+    /// byte-identical across Rust/TS/Swift.
     #[serde(default)]
-    pub progress: f32,
+    pub progress: u32,
 }
 
 /// A resume point for a video (offset within its duration).
